do some refacto
