+++ src/main/java/com/demo/medicrecruitment/job/providers/JpaJobOfferRepository.java
@@ -2,7 +2,7 @@ package com.demo.medicrecruitment.job.providers;
 
 import com.demo.medicrecruitment.job.domain.JobOfferDomain;
 import com.demo.medicrecruitment.job.infrastructure.JobOfferDomainRepository;
-import com.demo.medicrecruitment.job.mapper.domain.JobOfferDomainMapper;
+import com.demo.medicrecruitment.job.mapper.domain.JobOfferMapper;
 import com.demo.medicrecruitment.model.JobOffer;
 import com.demo.medicrecruitment.repository.JobOfferRepository;
 import org.springframework.stereotype.Component;
@@ -14,9 +14,9 @@ import java.util.List;
 class JpaJobOfferRepository implements JobOfferDomainRepository {
 
     private final JobOfferRepository jobOfferRepository;
-    private final JobOfferDomainMapper jobOfferDomainMapper;
+    private final JobOfferMapper jobOfferDomainMapper;
 
-    public JpaJobOfferRepository(JobOfferRepository jobOfferRepository, JobOfferDomainMapper jobOfferDomainMapper) {
+    public JpaJobOfferRepository(JobOfferRepository jobOfferRepository, JobOfferMapper jobOfferDomainMapper) {
         this.jobOfferRepository = jobOfferRepository;
         this.jobOfferDomainMapper = jobOfferDomainMapper;
     }
